==> 1c-coordinating_nodes_clothes.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import networkx as nx

from tqdm import tqdm

df = pd.read_csv("./clothes_result/nodes_clothes.csv")
# df = df.drop('reviewer_id', axis=1)
df = df.drop(['reviewer_id'], axis=1)
print(df.info())
df.head()

# edges = []

# for idx, row in df.iterrows():
#     edges.append((row['first_cycle'], row['second_cycle']))
#     edges.append((row['second_cycle'], row['category']))
    
# edges = np.array(edges)

edges = df[['first_cycle', 'second_cycle']].values
Q = nx.Graph()
Q.add_edges_from(edges)
coordinates = nx.spring_layout(Q)

coordinates

df_coor = pd.DataFrame\
            .from_dict(coordinates).T\
            .rename(columns={0: 'X', 1: 'Y'})\
            .reset_index()\
            .rename(columns={'index': 'NodeName'})

print(df_coor.info())
df_coor.head()

# Create connection bridge dataframe

temp_1 = df[['node_id', 'first_cycle']].rename(columns={'first_cycle': 'NodeName'})
temp_2 = df[['node_id', 'second_cycle']].rename(columns={'second_cycle':'NodeName'})

df_connection = pd.concat([temp_1, temp_2], axis=0).sort_values(by='node_id')

# # Create connection bridge
# edges = edges.reshape(-1)
# print(edges.shape)

# df_connector = pd.DataFrame(edges, columns=['NodeName'])

# array_idx = []
# for i in range(1, len(edges) + 1):
#     array_idx.append((i + 1) // 2)


# print(len(array_idx))
# df_connector['c_index'] = array_idx
# df_connector.head()

df_coor.to_csv("./clothes_result/coordinate_nodes_clothes.csv", index=False)
df_connection.to_csv("./clothes_result/connection_nodes_clothes.csv", index=False)
