==> 1d-summary_main_clothes.ipynb <==
import fasttext
model = fasttext.load_model('./util/cc.en.300.bin')

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import json
import re

from tqdm import tqdm

import util.normalization as norm
import util.model as models
import util.utility as util

# Opening JSON file
with open("./clothes_result/map_second_clothes.json") as json_file:
    map_second_keys = json.load(json_file)

with open("./clothes_result/map_category_clothes.json") as json_file:
    map_category = json.load(json_file)

with open("./clothes_result/list_first_cycle_clothes.json") as json_file:
    list_first_cycle = json.load(json_file)

df_base = pd.read_csv("./mcd_result/base_mcd_v2.csv")

print(df_base.info())
df_base.head()

def split_all(x):
    temp = []
    for e in x:
        if len(e) == 0:
            continue
        temp.append(tuple(e.split(", ")))
        
    return temp



df_base = df_base.fillna("")

df_base['aspect'] = df_base['aspect'].apply(lambda x: x.split(", "))
df_base['token_sentence'] = df_base['token_sentence'].apply(lambda x: x.split(", "))
df_base['token_lemma'] = df_base['token_lemma'].apply(lambda x: x.split(", "))
df_base['residu'] = df_base['residu'].apply(lambda x: x.split(", "))

df_base['key_score_aspect'] = df_base['key_score_aspect'].apply(lambda x: x.split("; ")).apply(lambda x: split_all(x))

print(df_base.info())
df_base.head()



df_main = df_base.copy()

df_main['token_sentence'] = df_main['token_sentence'].apply(lambda x: "\n".join("● " + item for item in x))

df_main.head()

def custom_text(x):
    temp = {}
    text = ""
    if not x:
        return text
    
    for s in x:
        category, key, sentiment = s

        if category not in temp:
            temp[category] = f"\t- {key} ({sentiment})"
        else:
            temp[category] = temp[category] + "\n" + f"\t- {key} ({sentiment})"



    categories = list(temp.keys())
    for category in categories:
        temp_text = f"● {category}:\n" + temp[category]

        text = text + temp_text + '\n'

    return text

df_main['information'] = df_main['key_score_aspect'].apply(lambda x: custom_text(x))

df_main = df_main.drop('key_score_aspect', axis=1)

columns = ['reviewer_id', 'review', 'summarize', 'token_sentence', 'information', 'rating']
df_main = df_main[columns]

df_main.to_csv('./clothes_result/main_clothes.csv', index=False)
