==> 1e-helper_analysis_mcd.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import re

from tqdm import tqdm

import util.normalization as norm
import util.model as models
import util.utility as util

df_base = pd.read_csv("./mcd_result/base_mcd.csv")

df_base['aspect'] = df_base['aspect'].apply(lambda x: x.split(", "))
df_base['token_sentence'] = df_base['token_sentence'].apply(lambda x: x.split(".\n"))
df_base['token_lemma'] = df_base['token_lemma'].apply(lambda x: x.split(".\n"))
print(df_base.info())
df_base.head()

# Locating Aspcet per Sentence

def bullet_aspect(x):
    texts, array = x
    temp = {}
    for text in texts:
        pattern = re.compile(rf"{text}")
        
        idx_result = [idx for idx, element in enumerate(array) if pattern.search(element)]
        
        i = 0
        while i < len(idx_result):
            curr_idx = idx_result[i]
            if curr_idx not in temp.keys():
                temp[curr_idx] = text
            else:
                temp[curr_idx] = temp[curr_idx] + f', {text}'
            i += 1
    return list(temp.values())

# df_base = df_base.explode('aspect')
df_base['temp_1'] = list(zip(df_base['aspect'].values, df_base['token_lemma'].values))

tqdm.pandas()
df_base['aspect'] = df_base['temp_1'].progress_apply(bullet_aspect)


df_base.drop('temp_1', axis=1, inplace=True)

df_base.head()

# Sentiment Score per Sentence

models.pattern_lexicon_model

def get_sentiment(x):
    temp = []
    for e in x:
        result = models.pattern_lexicon_model(e)
        temp.append(result)
    
    return temp


tqdm.pandas
df_base['sentiment_sentence'] = df_base['token_lemma'].apply(get_sentiment)

df_base.head()

df_base['token_lemma'].iloc[3]

df_base['aspect'].iloc[3]

df_base['aspect'] = df_base['aspect'].apply(lambda x: "\n ".join([f"• {item}" for item in x]))
df_base['sentiment_sentence'] = df_base['sentiment_sentence'].apply(lambda x: "\n ".join([f"• {item}" for item in x]))
df_base['token_sentence'] = df_base['token_sentence'].apply(lambda x: "\n ".join([f"• {item}" for item in x]))

df_base.head()

df_base[['reviewer_id', 'review', 'summarize', 
         'aspect', 'token_sentence', 'sentiment_sentence', 'rating']]\
        .to_csv("./mcd_result/helper_mcd.csv", index=False)
