==> 2-demo-aspect-sentiment-analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import json
import copy

from pattern.en import sentiment
from tqdm import tqdm

# Load JSON data from a file
with open('data-1.json', 'r') as json_file:
    data = json.load(json_file)

data

# Function for Sentiment Analysis using Pattern
def pattern_sentiment_analysis(text):
    # Get the sentiment polarity and subjectivity
    polarity, subjectivity = sentiment(text)
    
    # Determine the sentiment based on polarity
    if polarity > 0:
        sentiment_label = "Positive"
    elif polarity < 0:
        sentiment_label = "Negative"
    else:
        sentiment_label = "Neutral"
    
    # Return the results
    return {
        "sentiment": sentiment_label,
        "polarity": polarity,
        "subjectivity": subjectivity
    }

def sentiment_analysis(input_):
    # updated_data = input_.copy()
    updated_data = copy.deepcopy(input_)

    # updated_data = {}
    # Get dictionary level 1
    for i, dict_1 in updated_data.items():
        # Define new storage dictionary at level 1 base
        # new_dict_1 = {}
        # If the dictionary level 1 is empty, just determine sentiment overall text.
        if len(dict_1['aspect']) == 0:
            sentiment_ovl = pattern_sentiment_analysis(text)['sentiment']
            dict_1.update({'sentiment': sentiment_ovl})
            continue
        # Get dictionary level 2
        for aspect, dict_2 in dict_1['aspect'].items():
            # Define new storage dictionary at level 2 base
            new_dict = {}
            # Storage to store all text data for specific aspect
            texts = []
            # Get dictionary data at dictionary level 2
            for label, data in dict_2.items():
                # Define new_data storage
                new_data = []
                if len(data) > 0:
                    # Get through all data
                    for d in data:
                        text = d[-1]
                        texts.append(text)
                        # Sentiment analysis for text for specific keywords
                        result = pattern_sentiment_analysis(text)['sentiment']
                        # Update new data
                        new_data.append([d[0], d[1], result])
                # Update data dictonary level 2
                new_dict[label] = new_data
            # Update dictionary level 2
            dict_2.update(new_dict)
        # Concat all text into document, for sentiment analysis overall purpose.
        # NOTE: in case extracted text is empty use the raw review data.
        text = ''.join(texts) or dict_1['review']
        sentiment_ovl = pattern_sentiment_analysis(text)['sentiment']
        # Update new data
        dict_1.update({'sentiment': sentiment_ovl})
        # updated_data[i] = dict_1
    return updated_data

updated_data = sentiment_analysis(data)

updated_data

# testing = pd.DataFrame({'before': data.values(),
#                         'after': test.values()})

# testing.to_excel('fail.xlsx', index=False)

# df.to_excel('demo-sentiment.xlsx', index=False)

# # Save dictionary to a JSON file
# with open('data-2.json', 'w') as json_file:
#     json.dump(updated_data, json_file, indent=4)  # 'indent=4' makes the JSON pretty-printed
