==> 2.2-demo-topic-modeling.ipynb <==
# Utility: Preparation text

from contractions import CONTRACTION_MAP

##========== PREPARATION TEXT ===========##

# Contraction
def expand_contractions(sentence, contraction_mapping=CONTRACTION_MAP):
    """
    Expand the contractions in a sentence. For example don't => do not.
    
    Paramters:
    sentence (str): The input sentence to clean.
    contraction_mapping (dict): A dictionary for mapping contractions.
    
    
    Returns:
    str: The expanded contraction sentence.
    """
    
    contractions_pattern = re.compile('({})'.format('|'.join(contraction_mapping.keys())),
                                      flags=re.IGNORECASE|re.DOTALL)
    
    def expanded_match(contraction):
        """
        Filter for expanding the matched contraction.
        
        Parameters:
        contraction (str): The input of contraction
        
        Returns:
        str: The expanded contraction.
        """
        match = contraction.group(0)
        first_char = match[0]
        expanded_contraction = contraction_mapping.get(match) if contraction_mapping.get(match) else contraction_mapping.get(match.lower())
        
        expanded_contraction = first_char + expanded_contraction[1:]
        return expanded_contraction
    
    expanded_sentence = contractions_pattern.sub(expanded_match, sentence)
    return expanded_sentence


def remove_extra_spaces(sentence):
    # Use regex to replace multiple spaces with a single space
    return re.sub(r'\s+', ' ', sentence).strip()


def remove_non_ascii(text):
    """
    Remove all non-ASCII characters from the text.

    Parameters:
    text (str): The input text to clean.

    Returns:
    str: The cleaned text with only ASCII characters.
    
    """
    
    return ''.join([char for char in text if ord(char) < 128])

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import json
import copy
import re

from tqdm import tqdm

from nltk.corpus import stopwords

import spacy
import gensim
from pattern.en import sentiment

pd.set_option('display.max_colwidth', 2000)

# Load nlp model
nlp = spacy.load('en_core_web_sm')

# Load JSON data from a file
with open('data-2.json', 'r') as json_file:
    data = json.load(json_file)

data

def get_corpus(input_):
    corpus = []
    # Get dictionary level 1
    for i, dict_1 in input_.items():
        document = ''
        # Get dictionary level 2
        for aspect, dict_2 in dict_1['aspect'].items():
            # Get text from labels
            for label, data_ in dict_2.items():
                for d in data_:
                    document += d[1] + ' '
        corpus.append(document)

    return corpus

get_corpus(data)

corpus = get_corpus(data)

# Define the list of stopwords
stop_words = set(stopwords.words('english'))
    
# Preprocessing text
def preprocessing(text):
    text = remove_extra_spaces(text)
    text = expand_contractions(text)
    text = remove_non_ascii(text)

    # Get token of words
    doc = nlp(text)
    result = []
    for token in doc:
        t = token.lemma_.lower()

        if re.match(r'^[0-9\W]+$', t) or len(t) < 3 or t in stop_words:
            continue
        # If the token is adjective, noun, propn, or verb
        if token.pos_ in ['NOUN', 'PROPN']:
            result.append(t)
        elif token.pos_ in ['ADJ', 'VERB']:
            result.append(t)
        else:
            continue
    return result

# Create texts
texts = [preprocessing(document) for document in corpus]

# Create dictionary
dictionary = gensim.corpora.Dictionary(texts)


# Convert documents into Bag-of-words format
corpus_bow = [dictionary.doc2bow(text) for text in texts]

# Train the TF-IDF model
tfidf_model = gensim.models.TfidfModel(corpus_bow)

# Get corpus tfidf 
corpus_tfidf = tfidf_model[corpus_bow]

def topic_model_coherence_generator(corpus, texts, dictionary,
                                    start_topic_count=2, end_topic_count=10,
                                    step=1, cpus=1):
    models = []
    coherence_scores = []
    for topic_nums in tqdm(range(start_topic_count, end_topic_count+1, step)):
        lda_model = gensim.models.LdaModel(corpus=corpus, id2word=dictionary,
                                           chunksize=1740, alpha='auto',
                                           eta='auto', random_state=42,
                                           iterations=500, num_topics=topic_nums,
                                           passes=20, eval_every=None)

        cv_coherence_model_lda = gensim.models.CoherenceModel(model=lda_model,
                                                                     corpus=corpus,
                                                                     texts=texts,
                                                                     dictionary=dictionary,
                                                                     coherence='c_v')
        coherence_score = cv_coherence_model_lda.get_coherence()
        coherence_scores.append(coherence_score)
        models.append(lda_model)


    return models, coherence_scores

models, coherence_scores = topic_model_coherence_generator(corpus=corpus_tfidf,
                                                           texts=texts,
                                                           dictionary=dictionary)
opt_model = models[np.argmax(coherence_scores)]

# Calculate overall mean coherence score
topics_coherences = opt_model.top_topics(corpus_tfidf, topn=20)

coherence_scores

# Interpretation Result

# Visualize result: Topic with weights

topics_with_wts = [item[0] for item in topics_coherences]
print("LDA Topics with Weights")
print('='*50)
for idx, topic in enumerate(topics_with_wts):
  print(f'Topic {idx + 1}:')
  print([(term, round(wt, 3)) for wt, term in topic])
  print()

Explanation:
- Topic 1: Service Quality and Conflict Resolution.
> This topic likely centers around the overall quality of service and how customer issues are handled, particularly in resolving conflicts, complaints, or misunderstandings.
- Topic 2: Product Experience and Sales Interactions.
> This topic emphasizes product experiences and interactions with sales staff, with a particular focus on phones.
- Topic 3: Service Effectiveness and Customer Satisfaction.
> This topic likely reflects customer evaluations of service effectiveness.
- Topic 4: In-Store Experience and Helpfulness.
> This topic highlights in-store experiences, focusing on staff helpfulness, product selection, pricing, and store organization.

topics = [[(term, round(wt, 3))
              for term, wt in opt_model.show_topic(n, topn=20)]
          for n in range(0, opt_model.num_topics)
          ]

topic_df = pd.DataFrame([', '.join([term for term, wt in topic]) for topic in topics],
                       columns=['Term per Topic'],
                       index=[str(t) for t in range(1, opt_model.num_topics+1)])
topic_df

topic_json = topic_df.reset_index().rename({'index': 'topic', 'Term per Topic': 'term'}, axis=1).to_dict('records')

topic_json

# Interpreting result
tm_results = opt_model[corpus_tfidf]

# Corpus Topics
corpus_topics = [sorted(topics, key=lambda record: -record[1])[0]
                 for topics in tm_results]

corpus_topics[:5]

corpus_topic_df = pd.DataFrame()
corpus_topic_df['topic_dominant'] = [item[0]+1 for item in corpus_topics]
corpus_topic_df['topic_contribution'] = [round(item[1], 4) for item in corpus_topics]
corpus_topic_df['term'] = [topic_df.iloc[t[0]]['Term per Topic'] for t in corpus_topics]
corpus_topic_df['review'] = corpus

corpus_topic_df.head()

# Save result

# Update meta data
new_data = corpus_topic_df.drop('review', axis=1).to_dict('index')

def update_data(data_, new_data_):
    updated_data = copy.deepcopy(data_)

    for i, new in new_data_.items():
        updated_data[str(i)].update({'topic': new})

    return updated_data

save_data = update_data(data, new_data)

# Save dictionary to a JSON file
with open('data-3.json', 'w') as json_file:
    json.dump(save_data, json_file, indent=4)  # 'indent=4' makes the JSON pretty-printed


with open('data-topic.json', 'w') as json_file:
    json.dump(topic_json, json_file, indent=4)

# Save model
